cmake: compiler: gcc: Fix linker flags for RISC-V
The `-mabi`, `-march` and `-mcmodel` flags must be passed to the
linker in order to facilitate proper selection of the target multilib.

This commit also addresses the issue in which the `-mabi` and `-march`
flags are omitted for linking because `target_ld_options` evaluates the
two flags separately and they must be used in tandem in order for
compilation to succeed.

Note that this issue went unnoticed because other toolchain specs have
a hack in which the target library directory is directly added to the
library path using the `-L` option.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>