arch/x86/ia32: Include TLS GDT entry in count
When generating zephyr_pre1.elf, the number of GDT entries must be known so
that the memory layout of the executable is correct. This count didn't
include an entry for the thread local storage GDT entry when necessary,
making all addresses in _pre1 off by 8 bytes compared with the final elf
file.

Signed-off-by: Keith Packard <keithp@keithp.com>