scripts/utils: Build a list of "excluded" libc APIs
This script generates a list of exported symbols (functions, global
variables, static inlines and CPP names) which are present in picolibc but
not present in the minimal C library.

These symbols shall not be used by the "core Zephyr kernel"; the
construction of a script that detects use of the is left as an exercise to
the reader.

Signed-off-by: Keith Packard <keithp@keithp.com>