libc: fnmatch: Add to A.5 rule and common
fnmatch() is used by the shell, http server & coverage modules.
Given that quite a few subsystems need it,
let's add it to the C library common functions, and to the expected
functions from the C library in rule A.5,
instead of relaying on the corresponding POSIX subsystem option
being permanently enabled.

Several C libraries already implement it.
Those which do not, enable this common implementation.

Signed-off-by: Jukka Rissanen <jukka.rissanen@nordicsemi.no>